✨ feat(eventing): add codes for type identification and codes divipola, features in frontend

+++ services/catalog-service/src/main/java/com/msvanegasg/facturaelectronica/catalog/application/dto/DocumentTypeCommand.java
@@ -1,7 +1,7 @@
 package com.msvanegasg.facturaelectronica.catalog.application.dto;
 
 public record DocumentTypeCommand(
-        Long code,
+        Integer code,
         String name,
         String description) {
 }
